Rename middleware to middlewares
